Move flatten_key function to parser_json
